All can-import tests are passing again
This updates all of the can-import tests to use the new binding syntax.
One of the tests uses this method:

```handlebars
<can-import from='can/view/import/test/other.stache!' {^@value}='*other'
/>

{{{*other()}}}
```

While all of the others use the old syntax:

```handlebars
<can-import from='can/view/import/test/other.stache!' {^@value}='*other'
/>

{{> *other}}
```

So we know that both do work.

One change I had to make was to prevent can.Component from setting up
it's own bindings when can-import is used with `can-tag`. This is
because the bindings are performed in can/view/bindings instead.

Closes #2042 and #2044